Added Scripts and Prototype Scene NOT GOOD
I added multiple scripts to try to use have the player collect an item, which I was trying to make a key. I created a HUD for whenever the player collides with an item it will show in their Inventory HUD, but I wasn't successful in making it. I've been at it for about 4 hours, I'm wondering if you guys can take a look at it. It's under the scene called "PrototypeHub".